requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.13.1
matplotlib==3.9.0
river==0.20.1
PyYAML==6.0.2
joblib==1.4.2
requests==2.32.3
